+++ RELEASENOTES.md
@@ -31,6 +31,8 @@
         Use `Player.addListener` and `Player.Listener` instead.
     *   Remove `SimpleExoPlayer.addAudioListener`, `removeAudioListener` and
         `AudioListener`. Use `Player.addListener` and `Player.Listener` instead.
+    *   Remove `SimpleExoPlayer.addVideoListener`, `removeVideoListener` and
+        `VideoListener`. Use `Player.addListener` and `Player.Listener` instead.
 
 ### 2.15.0 (2021-08-10)